Insert 'prompt' parameters directly into functions
You are using these functions to modularize your code.  It seems like they are always just going to be called once. Even if a major change came around such as having dynamic transaction types in the prompt depending on the type of user, we would probably still want to encapsulate that prompt into the function and do everything there.  Also, it doesn't make much sense to have a parameter be static (not dynamic) in all the use cases where the function is being called.  That is smelly (https://en.wikipedia.org/wiki/Code_smell); that's why I refactored to cout the strings directly (no need to make them variables). And note lines 120-124, not sure if you knew you could just place string literals next to each other and they would automatically concatenate but I think that's pretty clean.

Oh, also you had a hanging function prototype that wasn't being used.. not sure if it is for future scope but I went ahead and removed that.